modified project structure; update README's; rename python files
