+++ cli/speech_generator_app.py
@@ -1,26 +1,14 @@
 import os
 import argparse
 from pathlib import Path
-from openai import OpenAI
-
-class TextToSpeechGenerator:
-    def __init__(self, api_key):
-        self.client = OpenAI(api_key=api_key)
-
-    def generate_speech(self, model, voice, input_text, output_path):
-        response = self.client.audio.speech.create(
-            model=model,
-            voice=voice,
-            input=input_text
-        )
-        response.stream_to_file(output_path)
+from ..src.speech_generator import TextToSpeechGenerator
 
 def parse_arguments():
     parser = argparse.ArgumentParser(description="Text-to-Speech Generator")
     parser.add_argument("--model", type=str, choices=["tts-1", "tts-1-hd"], default="tts-1", help="Model to use (default: tts-1)")
     parser.add_argument("--voice", type=str, choices=["alloy", "echo", "fable", "onyx", "nova", "shimmer"], help="Voice to use (default: alloy)")
     parser.add_argument("--input", type=str, required=True, help="Input text for generating speech")
-    parser.add_argument("--output", type=str, default="output/speech.mp3", help=f"Output file path for the generated speech (default: speech.mp3).  \n"
+    parser.add_argument("--output", type=str, default="outputs/speech.mp3", help=f"Output file path for the generated speech (default: speech.mp3).  \n"
                                                                                 f"Check list of supported output formats: https://platform.openai.com/docs/guides/text-to-speech")
     return parser.parse_args()
