--- samples/image-generator/requirements.txt
@@ -1,1 +1,0 @@
-openai==1.23.2
