+++ src/chat_bot.py
@@ -1,0 +1,16 @@
+from openai import OpenAI
+
+class ChatBot:
+    def __init__(self, api_key, model="gpt-3.5-turbo"):
+        self.client = OpenAI(api_key=api_key)
+        self.model = model
+
+    def chat(self, system_content, user_content):
+        completion = self.client.chat.completions.create(
+            model=self.model,
+            messages=[
+                {"role": "system", "content": system_content},
+                {"role": "user", "content": user_content}
+            ]
+        )
+        return completion.choices[0].message
