--- templates/secrets.template.cfg
@@ -1,1 +1,0 @@
-export OPENAI_API_KEY='PLACEHOLDER_OPENAI_API_KEY'
